refactored directory names +1
