+++ .idea/httpRequests/http-requests-log.http
@@ -1,0 +1,96 @@
+GET http://localhost:8080/notes
+
+<> 2019-01-05T072118.200.json
+
+###
+
+POST http://localhost:8080/notes
+Content-Type: application/json
+
+{
+  "author": "Richard",
+  "content": "hola5"
+}
+
+<> 2019-01-05T072114.200.json
+
+###
+
+GET http://localhost:8080/notes
+
+<> 2019-01-05T072108.200.json
+
+###
+
+GET http://localhost:8080/notes
+
+<> 2019-01-05T071938.200.json
+
+###
+
+POST http://localhost:8080/notes
+Content-Type: application/json
+
+{
+  "author": "Richard",
+  "content": "hola"
+}
+
+<> 2019-01-05T071936.200.json
+
+###
+
+GET http://localhost:8080/notes
+
+<> 2019-01-05T071930.200.json
+
+###
+
+POST http://localhost:8080/notes
+Content-Type: application/json
+
+{
+  "author": "Richard",
+  "content": "hola"
+}
+
+<> 2019-01-05T071928.200.json
+
+###
+
+GET http://localhost:8080/notes
+
+<> 2019-01-05T065531.200.json
+
+###
+
+POST http://localhost:8080/notes
+Content-Type: application/json
+
+{
+  "author": "Richard",
+  "content": "hola"
+}
+
+<> 2019-01-05T065528.200.json
+
+###
+
+GET http://localhost:8080/notes
+
+<> 2019-01-05T065524.200.json
+
+###
+
+POST http://localhost:8080/notes
+Content-Type: application/json
+
+{
+  "author": "Richard",
+  "content": "hola"
+}
+
+<> 2019-01-05T065520.200.json
+
+###
+
