+++ samples/sample-04/src/main/kotlin/com/example/domain/annotation/NoArg.kt
@@ -1,5 +1,5 @@
-package com.example.notesample.domain.annotation
+package com.example.domain.annotation
 
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.SOURCE)
-annotation class Open
+annotation class NoArg
