+++ samples/sample-04/src/main/kotlin/com/example/sampleApp.kt
@@ -1,4 +1,4 @@
-package com.example.notesample
+package com.example
 
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
