+++ samples/sample-04/src/main/kotlin/com/example/service/NoteService.kt
@@ -1,8 +1,8 @@
-package com.example.notesample.service
+package com.example.service
 
 import arrow.core.Option
 import arrow.data.NonEmptyList
-import com.example.notesample.domain.Note
+import com.example.domain.Note
 
 interface NoteService {
     fun getNotesByAuthor(author: String): Option<NonEmptyList<Note>>
